Updated fullcalendar.io files and scripts.
To add support for Google Calendars, I had to update all the Fullcalendar's
scripts. But when I did that, the website broke. So I had to delete everything
and work with the fullcalendar gem for rails, which worked like a charm.

Added all needed gems and deleted every other js file I had uploaded to github.
Updated every Require in the css and js application files and added an example
of Google Calendar API call. Also, recompiled our assets.